backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
Pillow>=10.4.0
httpx>=0.27.0
aiofiles>=24.1.0
pydantic>=2.10.0
python-dotenv>=1.0.0
cachetools>=5.5.0
